node: Fix gateway start ordering
Make sure that the handlers aren't added to the watchFactory until
after the gateway is fully initialized. This stops us dropping
any events.
