+++ go-controller/pkg/node/gateway.go
@@ -6,11 +6,13 @@ import (
 	"sync"
 
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/informer"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
 	util "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 	kapi "k8s.io/api/core/v1"
+	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog/v2"
 )
 
@@ -20,7 +22,7 @@ import (
 // are kept in sync
 type Gateway interface {
 	informer.ServiceAndEndpointsEventHandler
-	Init() error
+	Init(factory.NodeWatchFactory) error
 	Run(<-chan struct{}, *sync.WaitGroup)
 }
 
@@ -115,8 +117,43 @@ func (g *gateway) DeleteEndpoints(ep *kapi.Endpoints) {
 	}
 }
 
-func (g *gateway) Init() error {
-	return g.initFunc()
+func (g *gateway) Init(wf factory.NodeWatchFactory) error {
+	err := g.initFunc()
+	if err != nil {
+		return err
+	}
+	wf.AddServiceHandler(cache.ResourceEventHandlerFuncs{
+		AddFunc: func(obj interface{}) {
+			svc := obj.(*kapi.Service)
+			g.AddService(svc)
+		},
+		UpdateFunc: func(old, new interface{}) {
+			oldSvc := old.(*kapi.Service)
+			newSvc := new.(*kapi.Service)
+			g.UpdateService(oldSvc, newSvc)
+		},
+		DeleteFunc: func(obj interface{}) {
+			svc := obj.(*kapi.Service)
+			g.DeleteService(svc)
+		},
+	}, g.SyncServices)
+
+	wf.AddEndpointsHandler(cache.ResourceEventHandlerFuncs{
+		AddFunc: func(obj interface{}) {
+			ep := obj.(*kapi.Endpoints)
+			g.AddEndpoints(ep)
+		},
+		UpdateFunc: func(old, new interface{}) {
+			oldEp := old.(*kapi.Endpoints)
+			newEp := new.(*kapi.Endpoints)
+			g.UpdateEndpoints(oldEp, newEp)
+		},
+		DeleteFunc: func(obj interface{}) {
+			ep := obj.(*kapi.Endpoints)
+			g.DeleteEndpoints(ep)
+		},
+	}, nil)
+	return nil
 }
 
 func (g *gateway) Run(stopChan <-chan struct{}, wg *sync.WaitGroup) {
